move locale selector
